refactor(api): split MCP runtime modules
